drm/i915: Reduce number of register access during IVB+ interrupt handling
Register access is particularly obnoxious on Sandybridge and later due to
the extra work we must do around every read or write. The effect is
magnified on Haswell, as we have per-operation sanity checking
magnifying the number of reads and writes.

Interrupt handling is supposed to be fast, yet due to the sanity checks
around the register accesss it is not as fast as it could be. If we look
closer, most of the common register operations are reading values we
already know, and redundant flushes. Eliminate these by storing the
desired values rather than reading them back during the interrupt
routine.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>