intel-gtt: Report stolen_size as 0 when local memory is present
We don't have proper support for local memory elsewhere in the code,
and seeing as no-one seems to have ever seen a system which has local
memory, just report stolen_size as 0 when local memory is detected.

Add a big WARN to alert us if someone were to stumble on one of these
imaginary beasts.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>