drm/i915/sdvo: Fix LVDS connector status detection
There is an ordering issue currently in detecting the LVDS panel as
the determination depends upon whether there is a mode associated with
the connector. However, we don't associate modes with the connector
until we query get_modes() which is only done for a connected output.
The danger is that we then rely on the EDID for detecting the output,
which can be confused by attaching a VGA monitor and we erroneously
associate the EDID from the VGA output for the LVDS panel.

In order to provide accurate output detection, we need to query the LVDS
fixed-mode during detection and remove it if the SDVO output status
reports that it is ever decoupled.

Reported-by: Aitor Garcia <agarcia@tempel.es>
Reported-by: Wilhelm Hagemeister <hmr@igh-essen.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=63172
Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=71151
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>