drm/i915: Clarify the safety of the early unpin of old_fb->obj
Daniel simplified the modesetting code to push the common work performed
by each of the architecture specific routines higher into the caller. This
took me a while to be sure that it was safe in the event of a
modesetting failure - to be sure that the dangling pointer we left in
the registers would never be accessed. As it turns out, it is safe, as
even after the most convoluted error path, we always rewrite the address
registers with the currently pinned object before enabling the planes
and pipes. This patch just adds an assertion that the preconditions
Daniel stated are correct, and a comment to justify the dance.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>