drm/i915: Relax RPS contraints to allows setting minfreq on idle
When we idle, we set the GPU frequency to minimum. This should be a
safety blanket as the pcu on the GPU should be power gating itself
anyway. However, in order for us to do set the absolute minimum
frequency, we need to relax a few of our assertions that we do not
exceed the user limits.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>