drm/i915: Rename module parameters struct to i915_module
I want to use the short name i915 for the much more common practice of
refering to the device. Make the way clear by renaming the module
parameters. This has the added advantage of reinforcing that these are
global options and we should avoid intermixing device specific
information with them - i.e. when we santize (combine the module option
with device specific configuration) we should store that result locally
to the device.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>