drm/i915: s/seqno/request/ tracking inside objects
At the heart of this change is that the seqno is a too low level of an
abstraction to handle the growing complexities of command tracking, both
with the introduction of multiple command queues with execlists and the
potential for reordering with a scheduler. On top of the seqno we have
the request. Conceptually this is just a fence, but it also has
substantial bookkeeping of its own in order to track the context and
batch in flight, for example. It is the central structure upon which we
can extend with dependency tracking et al.

As regards the objects, they were using the seqno as a simple fence,
upon which is check or even wait upon for command completion. This patch
exchanges that seqno/ring pair with the request itself. For the
majority, lifetime of the request is ordered by how we retire objects
then requests. However, both the unlocked waits and probing elsewhere do
not tie into the normal request lifetimes and so we need to introduce a
kref. Extending the objects to use the request as the fence naturally
extends to segregating read/write fence tracking. This has significance
for it reduces the number of semaphores we need to emit, reducing the
likelihood of #54226, and improving performance overall.

v2: Rebase and split out the orthogonal tweaks.

A silly happened with this patch. It seemed to nullify our earlier
seqno-vs-interrupt w/a. I could not spot why, but gen6+ started to fail
with missed interrupts (a good test of our robustness handling). So I
ripped out the existing ACTHD read and replaced it with a RING_HEAD to
manually check whether the request is complete. That also had the nice
consequence of forcing __wait_request() to being the central arbiter of
request completion. Note that during testing, it was not enough to
re-enable the old workaround of keeping a forcewake reference whilst
waiting upon the interrupt+seqno.

The keener eyed reviewer will also spot that the reset_counter is moved
into the request simplifying __wait_request() callsites and reducing the
number of atomic reads by virtue of moving the check for a pending GPU
reset to the endpoints of GPU access.

v3: Implement the grand plan

Since execlist landed with its upside-down abstraction, unveil the power
of the request to remove all the duplication. To gain access to a ring,
you must allocate a request. To allocate a request you must specify the
context. Ergo all ring commands are carefully tracked by individual
requests (which demarcate a single complete transaction with the GPU) in
a known context (logical partitioning of the GPU with its own set of
registers and rings - which may be shared with other partitions for
backwards compatibility).

v4:

Tweak locking around execlist submission and request lists and
remove duplicated execlist code and the peppering of execlist
specific code throughout the core.

To simplify rebasing, I pulled in the s/ring/engine/ rename, it adds
a fair amount of noise of little significance and very easy to tune out.

The patch itself consists of 3 heavily intertwined parts:

0. Rename ring and engine variables to be consistent with their usage.
1. Change the ring access API to require the context under which we
   are operating. This generates a request which we use to build up a
   ring transaction. The request tracks required flushes and
   serialisation with both the GPU caches, other requests and the CPU.
2. Reorder initialisation such that we have a clearly defined context
   and engines for the early ring access on module load, resume and
   reset.
3. Convert the seqno tracking over to using requests (ala explicit
   fencing).

v5: Fix full-ppgtt (we have to allocate space on the default context for
the PDE instead of reusing the aliasing-ppgtt).

v6: Track activity on vma to fix ppgtt leaking.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Jesse Barnes <jbarnes@virtuousgeek.org>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Damien Lespiau <damien.lespiau@intel.com>
Cc: Oscar Mateo <oscar.mateo@intel.com>
Cc: Brad Volkin <bradley.d.volkin@intel.com>
Cc: "Kukanova, Svetlana" <svetlana.kukanova@intel.com>
Cc: Akash Goel <akash.goel@intel.com>
Cc: "Daniel, Thomas" <thomas.daniel@intel.com>
Cc: "Siluvery, Arun" <arun.siluvery@linux.intel.com>
Cc: John Harrison <John.C.Harrison@Intel.com>