drm/i915: Rearrange RPS frequency calculation
The goal is to refactor the Cherryview special casing to only be in a
single location. To do so we need to compute the desired adjustment and
then tweak it for Cherryview. This has the minor side-effect of making
sure we set adj to 0 if we directly update the target frequency.
Ensuring that the RPS constants are correct for Cherryview is outside
the scope of this function

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>