drm/i915: Make hangcheck logging more compact
When we detect a GPU hang, we emit a loud bang with the reason. But in
the process, we also log each ring that hangs and then lose that
information in the reason. Combine the two so that an accurate reason
why we triggered the GPU hang is logged in the error state and so that
we no longer need to emit the incremental hang detection.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>