drm/i915: Deflate error objects upon capture
When we capture the GPU error state, we allocate large amounts of memory
to preserve copies of the active objects on the GPU. We can compress the
copy in memory, and use an asci85 encoding when printing them out (to
avoid presenting binary data to unsuspecting catting of debugfs/procfs).

v2: Trailing newline and print all compressed objects correctly
v3: Use a ':' prefix to the line to mark compressed data for
intel_error_decode to detect more easily.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@intel.com>