drm/i915: Force a CS stall inside gen7 invalidate-caches
We use the invalidate-caches flush prior to doing a context switch and
so we rely on it being a command barrier. Frob the CS_STALL bit inside
the invalidate to make it a full barier before the context switch.

In the requests branch that started the hunt, this done as an explicit
engine->flush(COMMAND_BARRIER) before the MI_SET_CONTEXT.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=83677
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: stable@vger.kernel.org